requirements.txt
blinker==1.4
click==6.7
feedgen==0.9.0
Flask==1.0.0
Flask-HTTPAuth==3.2.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==20.1.0
itsdangerous==0.24
Jinja2==2.10.1
MarkupSafe==1.1.1
SQLAlchemy==1.3.0
Werkzeug==0.16.0
WTForms==2.1
bs4
requests
html5lib==1.0.1
PyPDF2
img2pdf
clipboard
bson
flask-socketio
eventlet==0.30.2
pymongo
flask-login
Pillow==8.1.0
urllib3==1.26.2
Flask-QRcode==3.0.0
numpy
imvickykumar999
adafruit-io
gevent
google-api-python-client
fpdf
youtube-search
pytube==10.8.5
pathlib
pytest-shutil
jsonify
pyttsx3
gTTS
xlsxwriter
secure-smtplib
moviepy==1.0.3
